requirements.txt
pyyaml
tqdm
opencv-python
imgaug
lpips
plyfile
tensorboard
sympy
numpy==1.23.5
pillow
matplotlib
bidict
ipdb
imageio
imageio-ffmpeg
open3d
timm
torchmetrics
kornia
roma
imgviz
protobuf
termcolor
segment_anything
# git+https://github.com/NVlabs/nvdiffrast.git
